Separation of functions and additional features.
Begun separating functions into separate files as well as added
additional functionality and features.
