+++ data/scripts/LogHelper.lua
