+++ data/scripts/ModBalancingEnabled.lua
@@ -1,0 +1,5 @@
+local modBalancingEnabled = GetModConfigData("MOD_BALANCING_ENABLED")
+
+return function() 
+	return modBalancingEnabled 
+end
