+++ data/scripts/ModifyCharacter.lua
@@ -9,15 +9,18 @@ function ModifyCharacter:addMode(inst, Key, Action)
 		if data.inst == ThePlayer then
 			if data.key == Key then
 				if TheWorld.ismastersim then
-					BufferedAction(inst, inst, ACTIONS[Action]):Do()
+					--BufferedAction(inst, inst, ACTIONS[Action]):Do()
+					inst:PushEvent("MCRAction")
 				else
-					SendRPCToServer(RPC.DoWidgetButtonAction, ACTIONS[Action].code, inst, ACTIONS[Action].mod_name)
+					--SendRPCToServer(RPC.DoWidgetButtonAction, ACTIONS[Action].code, inst, ACTIONS[Action].mod_name)
+					SendModRPCToServer( MOD_RPC.MCR[Action] )
 				end
 			end
 		end
 	end
-	inst:AddComponent("MCRkeyhandler")
-	inst:ListenForEvent("MCRkeypressed", OnKeyPressed)
+
+	inst:AddComponent("keyhandler")
+	inst:ListenForEvent("keypressed", OnKeyPressed)
 
 end
